Initial auth working.
Initial auth working with user email. Note that at present the redirect url in
the URL provided by Supabase needs to be updated to have /auth-demo/ prior to
the GET PATH.  This must be done manually before retrieving the URL (e.g. copy
the link out of the email sent from Supabase).