Add auth gated typed server data interface.
Add a set of libraries to interface with the TypeScript middleware client
in src/backend_client that are used for retrieving data, gating interface
elements on login, and for pagination.

Add Dashboard.tsx to demonstrate authn-gated cards, anon cards, and
pagination using the middleware API.

Details:

- The useApiData function takes typed structures directly from the client
  code and returns a typed API wrapper for use in a React context.
- The backend_client directory contains a client interface for the
  backend middleware (hosted today on GCP) and that uses a Supabase
  auth token for validation (validated by the middleware server).
- The ProtectedCard component gates card data requiring authn/authz
  behind a boiler plate interface with a login link.
- The DataCard component is used for retrieving and displaying server
  data and can be used within ProtectedCard elements for loading data
  after authn/authz.

See API_INTEGRATION_PATTERNS.md for more technical detail on these
changes.

Notes regarding end-to-end schmea:

The OpenAPI specification is generated directly from the middleware-server
repo's generate_openapi.py script. The models included in OpenAPI are created
by the src/sqlacodegen.sh script in the middleware server repo and are specified
in the FastAPI request and response types in code.

Accordingly, a schema change in the postgres database, such as the
addition of a new column, has a primary source of truth in postgres,
and can be propagated through to the frontend in the following sequence:

  backend: sqlacodegen.sh
  backend: generate_openapi.py
  frontend: npx @hey-api/openapi-ts -i openapi.json   -o src/backend_client

The supabase auto-generated OpenAPI schemas are NOT USED today, though
there are strong arguments to use them as the primary client API if
and when we commit to Supabase. As of this writing Supabase is used only
for user account management and as a Postgres database.